fix numeric types

+++ MyLearn/MyLearn/InputModels/CustomTopStudent.cs
@@ -2,10 +2,10 @@
 {
     public class CustomTopStudent
     {
-        public int NumberOfTopStudents { get; set; }
-        public int CourseAvgWeight { get; set; }
-        public int CourseSuccessRateWeight { get; set; }
-        public int ProjectAvgWeight { get; set; }
-        public int ProjectSuccessRateWeight { get; set; }
+        public float NumberOfTopStudents { get; set; }
+        public float CourseAvgWeight { get; set; }
+        public float CourseSuccessRateWeight { get; set; }
+        public float ProjectAvgWeight { get; set; }
+        public float ProjectSuccessRateWeight { get; set; }
     }
 }
