+++ MyLearn/MyLearn/InputModels/NewCourse.cs
@@ -8,8 +8,8 @@ namespace MyLearn.InputModels
         public string CourseDescription { get; set; }
         public string ProfUserId { get; set; }
         public string UniversityId { get; set; }
-        public int Group { get; set; }
-        public int MinGrade { get; set; }
+        public float Group { get; set; }
+        public float MinGrade { get; set; }
         public List<Badge> Badges { get; set; }
     }
 }
