+++ MyLearn/MyLearn/InputModels/NewJobOffer.cs
@@ -10,6 +10,6 @@ namespace MyLearn.InputModels
         public string StartDate { get; set; }
         public string EndDate { get; set; }
         public string Description { get; set; }
-        public int Budget { get; set; }
+        public float Budget { get; set; }
     }
 }
