+++ MyLearn/MyLearn/Models/Comment.cs
@@ -7,7 +7,7 @@ namespace MyLearn.Models
         public string CommentId { get; set; }
         public string ParentId { get; set; }
         public string Date { get; set; }
-        public int IsFromStudent { get; set; }
+        public float IsFromStudent { get; set; }
         public string CommentContent { get; set; }
         public List<Comment> NestedComments { get; set; }
     }
