+++ MyLearn/MyLearn/Models/Course.cs
@@ -6,10 +6,10 @@ namespace MyLearn.Models
     {
         public string CourseName { get; set; }
         public string UniversityId { get; set; }
-        public int MinGrade { get; set; }
+        public float MinGrade { get; set; }
         public string CourseId { get; set; }
         public string CourseDescription { get; set; }
-        public int Group { get; set; }
+        public float Group { get; set; }
         public List<StudentInCourse> Students { get; set; }
     }
 }
