+++ MyLearn/MyLearn/Models/JobOffer.cs
@@ -10,8 +10,8 @@ namespace MyLearn.Models
         public string StartDate { get; set; }
         public string EndDate { get; set; }
         public string Description { get; set; }
-        public int Budget { get; set; }
-        public int State { get; set; }
+        public float Budget { get; set; }
+        public float State { get; set; }
         public string StateDescription { get; set; }
     }
 }
